+++ playroom/snippets.tsx
@@ -228,7 +228,7 @@ const listSnippetsAvatar: Array<Snippet> = [
                   <Avatar
                     size={40}
                     initials="ML"
-                    src="https://api.lorem.space/image/face?w=200&h=200"
+                    src="https://source.unsplash.com/600x600/?face"
                   />
                 }
                 title="Title"
@@ -290,7 +290,7 @@ const listRowSnippets: Array<Snippet> = ['Row', 'BoxedRow'].flatMap((rowName) =>
         name: `${rowName} (image)`,
         code: `
         <${rowName}
-            asset={<Image src="https://api.lorem.space/image/album?w=300&h=300" height={80} aspectRatio="1:1" />}
+            asset={<Image src="https://source.unsplash.com/900x900/?landscape" height={80} aspectRatio="1:1" />}
             title="Title"
             description="Description"
             onPress={() => {}}
@@ -719,12 +719,32 @@ const cardSnippets: Array<Snippet> = [
           pretitle="Pretitle"
           title="Title"
           description="Description"
-          backgroundImage="https://api.lorem.space/image/watch?w=700&h=700"
+          backgroundImage="https://source.unsplash.com/900x900/?landscape"
           button={
             <ButtonPrimary small href="https://google.com">
               Action
             </ButtonPrimary>
           }
+          onClose={() => {}}
+          actions={[
+            {
+              Icon: IconLightningRegular,
+              onPress: () => {},
+              label: "Lightning",
+            },
+          ]}
+        />`,
+    },
+    {
+        group: 'Cards',
+        name: 'PosterCard',
+        code: `
+        <PosterCard
+          headline={<Tag type="promo">Headline</Tag>}
+          pretitle="Pretitle"
+          title="Title"
+          description="Description"
+          backgroundImage="https://api.lorem.space/image/watch?w=700&h=700"
           onClose={() => {}}
           actions={[
             {
@@ -931,7 +951,7 @@ const exampleScreens: Array<Snippet> = [
               <Avatar
                 size={isDesktopOrBigger ? 32 : 24}
                 initials="ML"
-                src="https://api.lorem.space/image/face?w=200&h=200"
+                src="https://source.unsplash.com/600x600/?face"
               />
               {isDesktopOrBigger && "María López Serrano"}
             </NavigationBarAction>
@@ -1204,7 +1224,7 @@ const navigationBarSnippets = [
         <Avatar
           size={isDesktopOrBigger ? 32 : 24}
           initials="ML"
-          src="https://api.lorem.space/image/face?w=200&h=200"
+          src="https://source.unsplash.com/600x600/?face"
         />
         {isDesktopOrBigger && "María López Serrano"}
       </NavigationBarAction>
@@ -1280,7 +1300,7 @@ const carouselSnippets = [
       headline={<Tag type="promo">Headline</Tag>}
       title={'Card ' + idx}
       description="Description"
-      media={<Image src="https://api.lorem.space/image/watch?w=700&h=700" aspectRatio="16:9" />}
+      media={<Image src="https://source.unsplash.com/1600x900/?watch" aspectRatio="16:9" />}
       button={
         <ButtonPrimary small onPress={() => {}}>
           Action
@@ -1298,8 +1318,8 @@ const carouselSnippets = [
 <Slideshow
   withBullets
   items={[
-    <Image src="https://api.lorem.space/image/furniture?w=1500&h=1500&hash=8B7BCDC2" aspectRatio="16:9" />,
-    <Image src="https://api.lorem.space/image/furniture?w=1500&h=1500&hash=500B67FB" aspectRatio="16:9" />,
+    <Image src="https://source.unsplash.com/1600x900/?furniture" aspectRatio="16:9" />,
+    <Image src="https://source.unsplash.com/1600x900/?furniture" aspectRatio="16:9" />,
   ]}
 />`,
     },
@@ -1336,12 +1356,12 @@ const avatarSnippets = [
     {
         group: 'Avatar',
         name: 'Avatar with badge',
-        code: `<Avatar size={64} src="https://api.lorem.space/image/face?w=200&h=200" badge />`,
+        code: `<Avatar size={64} src="https://source.unsplash.com/600x600/?face" badge />`,
     },
     {
         group: 'Avatar',
         name: 'Avatar with numeric badge',
-        code: `<Avatar size={64} src="https://api.lorem.space/image/face?w=200&h=200" badge={5} />`,
+        code: `<Avatar size={64} src="https://source.unsplash.com/600x600/?face" badge={5} />`,
     },
     {
         group: 'Avatar',
@@ -1574,7 +1594,7 @@ const heroSnippets = [
                 background="default"
                 media={
                   <Image
-                    src="https://api.lorem.space/image/furniture?w=1500&h=1500&hash=8B7BCDC2"
+                    src="https://source.unsplash.com/1600x900/?furniture"
                     aspectRatio="16:9"
                   />
                 }
@@ -1592,6 +1612,55 @@ const heroSnippets = [
           />
 
       `,
+    },
+];
+
+const loaderSnippets = [
+    {
+        name: 'Loader with url',
+        code: `
+          <Loader
+          load="https://randomuser.me/api/?results=20&seed=patata&nat=es"
+          render={(data) => (
+            <RowList>
+              {data.results.map((p) => (
+                <Row
+                  asset={<Avatar size={40} src={p.picture.medium} />}
+                  title={p.name.first}
+                  description={p.email}
+                />
+              ))}
+            </RowList>
+          )}
+          renderLoading={() => <Spinner />}
+          renderError={() => <Snackbar message="An error occurred :(" type="CRITICAL" />}
+          />
+        `,
+        group: 'Loader',
+    },
+    {
+        name: 'Loader with function',
+        code: `
+          <Loader
+            load={async () => {
+                return fetch('https://randomuser.me/api/?results=20&seed=patata&nat=es').then((r) => r.json());
+            }}
+            render={(data) => (
+                <RowList>
+                    {data.results.map((p) => (
+                        <Row
+                            asset={<Avatar size={40} src={p.picture.medium} />}
+                            title={p.name.first}
+                            description={p.email}
+                        />
+                    ))}
+                </RowList>
+            )}
+            renderLoading={() => <Spinner />}
+            renderError={() => <Snackbar message="An error occurred :(" type="CRITICAL" />}
+          />
+        `,
+        group: 'Loader',
     },
 ];
 
@@ -1727,4 +1796,5 @@ export default [
     ...tooltipSnippets,
     ...popoverSnippets,
     ...heroSnippets,
+    ...loaderSnippets,
 ].sort((s1, s2) => s1.group.localeCompare(s2.group)) as Array<Snippet>;