+++ src/__stories__/highlighted-card-carousel-story.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import {Box, ResponsiveLayout, HighlightedCard, Carousel, ButtonPrimary} from '..';
+import takingPicturePortraitImg from './images/taking-picture-portrait.jpg';
 
 export default {
     title: 'Components/Carousels/Highlighted card carousel',
@@ -18,12 +19,12 @@ export const Default: StoryComponent<Args> = () => {
                         <HighlightedCard
                             title="Title1"
                             description="This is the best offer for you. Please check it out!"
-                            imageUrl="https://images.unsplash.com/photo-1557180295-76eee20ae8aa?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1160&q=80"
+                            imageUrl={takingPicturePortraitImg}
                         />,
                         <HighlightedCard
                             title="Title2"
                             description="This is the best offer for you. Please check it out!"
-                            imageUrl="https://images.unsplash.com/photo-1557180295-76eee20ae8aa?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1160&q=80"
+                            imageUrl={takingPicturePortraitImg}
                             button={<ButtonPrimary onPress={() => {}}>Action</ButtonPrimary>}
                             onClose={() => {}}
                         />,